ref(util,pkg/ssh,cmd/keys.go): move util package to pkg/ssh
