+++ pkg/ssh/doc.go
@@ -1,0 +1,3 @@
+// Package ssh contains SSH related utility functions that are used by both the CLI binary
+// and the client SDK
+package ssh
