--- util/doc.go
@@ -1,3 +1,0 @@
-// Package util contains various utility functions that are used by both the CLI binary
-// and the client SDK
-package util
